ci: route crawler tests through nkust_crawler package
Issue #331's '方案 B' (scheduled live monitor) was workflow-passing but
test-failing because flutter test in crawler-monitor.yml could not
resolve native_dio_adapter / syncfusion_flutter_pdf / image_picker
plugins on a Linux runner. Now that the crawler is a pure-Dart package,
that whole class of failure goes away.

ci.yml — Analyze & Test job:
- watch packages/** on push/PR
- analyze + test the sub-package via dart analyze / dart test inside
  packages/nkust_crawler (no Flutter SDK needed)
- drop '|| true' from 'flutter test' so main-app fixture test failures
  block PR merge again (Issue #331 方案 A is now actually enforced)

crawler-monitor.yml:
- replace flutter test with dart test -P live against the package
- map existing NKUST_USERNAME / NKUST_PASSWORD secrets to the package's
  expected NKUST_USER / NKUST_PASS env vars
- swap subosito/flutter-action for dart-lang/setup-dart

New Health Check suite in packages/nkust_crawler/test/live/ covers the
6 endpoints the old test had (Oosaf supersedes the retired
leave.nkust.edu.tw DNS record). Test names still start with 'Health Check'
so the classify-failures step can distinguish server-down from
parser-drift.

Delete test/crawler_monitor_test.dart — its coverage is now in
packages/nkust_crawler/test/live/{health_check,authenticated}_live_test.dart.
